# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API документашка'
  version: 1.0.0
paths:
...
  /api/users:
    get:
      tags:
        - User
      summary: 'INDEX - Список пользователей'
      operationId: c457726701591d1183b53aa71fc13441
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: 'Семён Семёныч'
                        email:
                          type: string
                          example: example@mail.org
                        address:
                          type: string
                          example: 'ул. Кукурузная, д. 35'
                        phone:
                          type: string
                          example: '+79999999999'
                        orders:
                          type: integer
                          example: 3
                        orders_total:
                          type: integer
                          example: 161100
                        permissions:
                          type: arr
                          example: '[...]'
                      type: object
                  links:
                    type: obj
                    example: '{...}'
                  meta:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
    post:
      tags:
        - User
      summary: 'STORE - Регистрация пользователя и выдача пары токенов'
      operationId: 592819a0265360b2014512d6dbfaf0e7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Василий Ложкин'
                email:
                  type: string
                  example: vasya@mail.org
                password:
                  type: string
                  example: qwerty
                password_confirmation:
                  type: string
                  example: qwerty
              type: object
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  access:
                    type: string
                    example: abcdefg...
                  access_exp:
                    type: string
                    example: 1535153
                  user:
                    type: obj
                    example: '{id, ...}'
                type: object
        422:
          description: 'Валидация не пройдена'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Поле email обязательно. (and 1 more error)'
                  errors:
                    properties:
                      email:
                        type: arr
                        example: "['Поле email обязательно.']"
                      password:
                        type: arr
                        example: "['Поле пароль обязательно.']"
                    type: object
                type: object
  '/api/users/{user}':
    get:
      tags:
        - User
      summary: 'SHOW - Информация о пользователе'
      operationId: d3a14968f8b7a072620c75c7c78ed6f8
      parameters:
        -
          name: user
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Семён Семёныч'
                      email:
                        type: string
                        example: example@mail.org
                      address:
                        type: string
                        example: 'ул. Кукурузная, д. 35'
                      phone:
                        type: string
                        example: '+79999999999'
                      orders:
                        type: integer
                        example: 3
                      orders_total:
                        type: integer
                        example: 161100
                      permissions:
                        type: arr
                        example: '[...]'
                      created_at:
                        type: string
                        example: '2024-04-09T16:08:39.000000Z'
                    type: object
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
  '/api/users/{id}/password':
    patch:
      tags:
        - User
      summary: 'changePassword - Изменение пароля'
      operationId: ae7266b3a8dea94874e25be9ce2e1299
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                old_password:
                  type: string
                  example: qwerty
                new_password:
                  type: string
                  example: '12345'
                new_password_confirmation:
                  type: string
                  example: '12345'
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Пароль изменён.'
                  user:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Пользователь не найден, либо нет прав.'
        422:
          description: 'Валидация не пройдена'
      security:
        -
          jwt: []
  '/api/users/{id}/address':
    patch:
      tags:
        - User
      summary: 'changeAddress - Изменение адреса'
      operationId: 787598a620bf9a838cc08b02fa93e906
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                address:
                  type: string
                  example: 'г. Череповец, ул. Кирпичная, д. 3, кв. 1'
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Адрес изменён.'
                  user:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Пользователь не найден, либо нет прав.'
        422:
          description: 'Валидация не пройдена'
      security:
        -
          jwt: []
  '/api/users/{id}/profile':
    patch:
      tags:
        - User
      summary: 'changeProfile - Изменение имени и телефона'
      operationId: 312bc0e86a98f0a75e83afdd8bdc812c
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Семён семёныч'
                phone:
                  type: string
                  example: '+70123456789'
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Имя и телефон изменены.'
                  user:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Пользователь не найден, либо нет прав.'
        422:
          description: 'Валидация не пройдена'
      security:
        -
          jwt: []
  '/api/users/{id}/email':
    patch:
      tags:
        - User
      summary: 'changeEmail - Изменение email'
      description: 'password - действующий пароль для подтверждения операции'
      operationId: f4f28a6b2cf871503d670fc0992d943e
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  example: example@mail.org
                password:
                  type: string
                  example: qwerty
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Email изменён.'
                  user:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Пользователь не найден, либо нет прав.'
        422:
          description: 'Валидация не пройдена, либо неверный пароль.'
      security:
        -
          jwt: []
  '/api/users/{id}':
    delete:
      tags:
        - User
      summary: 'DESTROY - Удаление пользователя'
      operationId: fa56cffde745d3f152f95cbacd936c0b
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Пользователь удалён.'
                  user:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Нет прав, либо пользователь не существует'
      security:
        -
          jwt: []
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer